rename engines package to engine to reflect that this is not about different engines, but about the engine framework
