+++ Extensions/Pipelining/src/main/scala/rescala/pipelining/PipelineEngine.scala
@@ -1,6 +1,6 @@
 package rescala.pipelining
 
-import rescala.engines.PlanImpl
+import rescala.engine.PlanImpl
 
 /**
  * @author moritzlichter
