+++ Main/jvm/src/main/scala/rescala/package.scala
@@ -1,4 +1,4 @@
-import rescala.engines.EngineImpl
+import rescala.engine.EngineImpl
 import rescala.parrp.{Backoff, ParRP}
 
 package object rescala extends EngineImpl[ParRP, ParRP]("ParRP", (_, prior) => new ParRP(new Backoff, prior)) {
