+++ Main/shared/src/main/scala/rescala/engine/Engine.scala
@@ -1,4 +1,4 @@
-package rescala.engines
+package rescala.engine
 
 import rescala.graph.Struct
 import rescala.macros.ReactiveMacros
@@ -26,7 +26,7 @@ trait Engine[S <: Struct, +TTurn <: Turn[S]] {
   final type Var[A] = reactives.Var[A, S]
   final type Evt[A] = reactives.Evt[A, S]
   final type Turn = propagation.Turn[S]
-  final type Ticket = rescala.engines.TurnSource[S]
+  final type Ticket = rescala.engine.TurnSource[S]
   final type Reactive = rescala.graph.Reactive[S]
   final def Evt[A](): Evt[A] = reactives.Evt[A, S]()
