+++ Main/shared/src/main/scala/rescala/reactives/Events.scala
@@ -1,6 +1,6 @@
 package rescala.reactives
 
-import rescala.engines.TurnSource
+import rescala.engine.TurnSource
 import rescala.graph.Pulse.NoChange
 import rescala.graph._
 import rescala.propagation.Turn
