+++ Main/shared/src/main/scala/rescala/reactives/Signal.scala
@@ -1,6 +1,6 @@
 package rescala.reactives
 
-import rescala.engines.{Engine, TurnSource}
+import rescala.engine.{Engine, TurnSource}
 import rescala.graph.{Pulse, Stateful, Struct}
 import rescala.propagation.Turn
 import rescala.reactives.RExceptions.EmptySignalControlThrowable
