+++ Research/Meta/src/main/scala/rescala/meta/DataFlowNode.scala
@@ -1,6 +1,6 @@
 package rescala.meta
 
-import rescala.engines.{Engine, TurnSource}
+import rescala.engine.{Engine, TurnSource}
 import rescala.graph.{Pulse, Pulsing, Struct}
 import rescala.propagation.Turn
 import rescala.reactives._
