+++ Research/Meta/src/main/scala/rescala/meta/DataFlowRef.scala
@@ -1,6 +1,6 @@
 package rescala.meta
 
-import rescala.engines.TurnSource
+import rescala.engine.TurnSource
 import rescala.graph.Struct
 import rescala.reactives._
