+++ Research/Microbenchmarks/src/main/scala/benchmarks/STMBank/BankAccounts.scala
@@ -6,7 +6,7 @@ import java.util.concurrent.{ThreadLocalRandom, TimeUnit}
 import benchmarks.EngineParam
 import org.openjdk.jmh.annotations._
 import org.openjdk.jmh.infra.{BenchmarkParams, Blackhole}
-import rescala.engines.Engine
+import rescala.engine.Engine
 import rescala.graph.Struct
 import rescala.propagation.Turn
 import rescala.reactives._
