+++ Research/Microbenchmarks/src/main/scala/benchmarks/chatserver/ChatServer.scala
@@ -1,6 +1,6 @@
 package benchmarks.chatserver
 
-import rescala.engines.Engine
+import rescala.engine.Engine
 import rescala.graph.Struct
 import rescala.propagation.Turn
