+++ Research/Microbenchmarks/src/main/scala/benchmarks/simple/SingleVar.scala
@@ -8,7 +8,7 @@ import org.openjdk.jmh.annotations._
 import org.openjdk.jmh.infra.BenchmarkParams
 import rescala.Engines
 import rescala.benchmarkutil.BenchmarkUtil
-import rescala.engines.Engine
+import rescala.engine.Engine
 import rescala.propagation.Turn
 import rescala.reactives.Var
