+++ Research/Microbenchmarks/src/main/scala/benchmarks/simple/TurnCreation.scala
@@ -5,7 +5,7 @@ import java.util.concurrent.TimeUnit
 import benchmarks.{EngineParam, Workload}
 import org.openjdk.jmh.annotations._
 import org.openjdk.jmh.infra.BenchmarkParams
-import rescala.engines.Engine
+import rescala.engine.Engine
 import rescala.propagation.Turn
 
 @BenchmarkMode(Array(Mode.Throughput))
