+++ Tests/jvm/src/test/scala/tests/rescala/concurrency/philosophers/PhilosopherTable.scala
@@ -2,7 +2,7 @@ package tests.rescala.concurrency.philosophers
 
 import java.util.concurrent.atomic.AtomicInteger
 
-import rescala.engines.Engine
+import rescala.engine.Engine
 import rescala.graph.Globals.named
 import rescala.graph.Struct
 import rescala.parrp.Backoff
