+++ Tests/shared/src/test/scala/tests/rescala/TurnSourceTest.scala
@@ -6,7 +6,7 @@ class TurnSourceTest extends RETests {
 
   /* this test uses some shady planned()(identity) to get the turn object out of the transaction
    * you should not do this. */
-  def getTurn[S2 <: rescala.graph.Struct](implicit engine: rescala.engines.Engine[S2, rescala.propagation.Turn[S2]]): rescala.propagation.Turn[S2] = engine.plan()(identity)
+  def getTurn[S2 <: rescala.graph.Struct](implicit engine: rescala.engine.Engine[S2, rescala.propagation.Turn[S2]]): rescala.propagation.Turn[S2] = engine.plan()(identity)
 
   allEngines("none Dynamic No Implicit") { engine => import engine._
     assert(implicitly[Ticket].self === Right(engine))
